+++ xarray/core/duck_array_ops.py
@@ -30,11 +30,11 @@ from numpy import (  # noqa
     transpose,
     unravel_index,
 )
-from numpy.lib.stride_tricks import sliding_window_view  # noqa
+from numpy.ma import masked_invalid  # noqa
 from packaging.version import Version
 from pandas.api.types import is_extension_array_dtype
 
-from xarray.core import dask_array_ops, dtypes, nputils
+from xarray.core import dask_array_compat, dask_array_ops, dtypes, nputils
 from xarray.core.options import OPTIONS
 from xarray.core.utils import is_duck_array, is_duck_dask_array, module_available
 from xarray.namedarray import pycompat
@@ -92,11 +92,12 @@ def _dask_or_eager_func(
     name,
     eager_module=np,
     dask_module="dask.array",
+    dask_only_kwargs=tuple(),
 ):
     """Create a function that dispatches to dask for dask array inputs."""
 
     def f(*args, **kwargs):
-        if any(is_duck_dask_array(a) for a in args):
+        if dask_available and any(is_duck_dask_array(a) for a in args):
             mod = (
                 import_module(dask_module)
                 if isinstance(dask_module, str)
@@ -105,6 +106,8 @@ def _dask_or_eager_func(
             wrapped = getattr(mod, name)
         else:
             wrapped = getattr(eager_module, name)
+            for kwarg in dask_only_kwargs:
+                kwargs.pop(kwarg)
         return wrapped(*args, **kwargs)
 
     return f
@@ -121,6 +124,15 @@ def fail_on_dask_array_input(values, msg=None, func_name=None):
 
 # Requires special-casing because pandas won't automatically dispatch to dask.isnull via NEP-18
 pandas_isnull = _dask_or_eager_func("isnull", eager_module=pd, dask_module="dask.array")
+
+# sliding_window_view will not dispatch arbitrary kwargs (automatic_rechunk),
+# so we need to hand-code this.
+sliding_window_view = _dask_or_eager_func(
+    "sliding_window_view",
+    eager_module=np.lib.stride_tricks,
+    dask_module=dask_array_compat,
+    dask_only_kwargs=("automatic_rechunk",),
+)
 
 
 def round(array):
@@ -168,12 +180,6 @@ def isnull(data):
 
 def notnull(data):
     return ~isnull(data)
-
-
-# TODO replace with simply np.ma.masked_invalid once numpy/numpy#16022 is fixed
-masked_invalid = _dask_or_eager_func(
-    "masked_invalid", eager_module=np.ma, dask_module="dask.array.ma"
-)
 
 
 def trapz(y, x, axis):